Apply Rubocop
Current Rubocop does not support Ruby 2.2 and so I've decided to drop support for it here. This is unlikely to be an issue (how many projects are still using 2.2 and will want to use this gem?) but I feel that if the tool chain is dropping it, I should probably follow suit.